// link2rail/projects/ekp/src/app/invoice/api/generated/model/invoice-summary-response.ts
/**
 * Invoice API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InvoiceSummary } from './invoice-summary';


export interface InvoiceSummaryResponse { 
    /**
     * offset within the result set, as send in the request
     */
    offset: number;
    /**
     * the maximum number of returned items
     */
    limit: number;
    /**
     * Total number of elements
     */
    total: number;
    summaries?: Array<InvoiceSummary>;
}